chore: removed old DatabaseManager from Catalog Next
