+++ plugins/catalog-backend/src/next/NextCatalogBuilder.ts
@@ -14,11 +14,7 @@
  * limitations under the License.
  */
 
-import {
-  PluginDatabaseManager,
-  resolvePackagePath,
-  UrlReader,
-} from '@backstage/backend-common';
+import { PluginDatabaseManager, UrlReader } from '@backstage/backend-common';
 import {
   DefaultNamespaceEntityPolicy,
   EntityPolicies,
@@ -70,6 +66,7 @@ import {
 } from '../next/types';
 import { ConfigLocationEntityProvider } from './ConfigLocationEntityProvider';
 import { DefaultProcessingDatabase } from './database/DefaultProcessingDatabase';
+import { applyDatabaseMigrations } from './database/migrations';
 import { DefaultCatalogProcessingEngine } from './DefaultCatalogProcessingEngine';
 import { DefaultLocationService } from './DefaultLocationService';
 import { DefaultLocationStore } from './DefaultLocationStore';
@@ -304,12 +301,7 @@ export class NextCatalogBuilder {
     const parser = this.parser || defaultEntityDataParser;
 
     const dbClient = await database.getClient();
-    await dbClient.migrate.latest({
-      directory: resolvePackagePath(
-        '@backstage/plugin-catalog-backend',
-        'migrations',
-      ),
-    });
+    await applyDatabaseMigrations(dbClient);
 
     const db = new CommonDatabase(dbClient, logger);
